+++ packages/terra-clinical-detail-view/src/terra-dev-site/test/clinical-detail-view/DetailView/DetailViewDivided.test.jsx
@@ -19,7 +19,7 @@ const DetailViewDivided = () => (
       <div className={cx('graph-content')}>
         This is where any visualizations would go
       </div>
-)}
+    )}
     accessory={accessoryElement}
     details={[
       (