+++ packages/terra-clinical-detail-view/src/terra-dev-site/test/clinical-detail-view/DetailView/DetailViewDividedSmallerTitles.test.jsx
@@ -1,11 +1,21 @@
 import React from 'react';
 import classNames from 'classnames/bind';
+import IconCritical from 'terra-icon/lib/icon/IconCritical';
 import DetailView from '../../../../DetailView';
 import styles from './DetailViewCommon.test.module.scss';
 
 const cx = classNames.bind(styles);
 
-const item = (<p>A Detail List Item</p>);
+const item = (
+  <div className={cx('indicator-container')}>
+    <div className={cx('icon')}>
+      <IconCritical a11yLabel="Critical Icon" />
+      <div className={cx('text')}>
+        Immediate Priority
+      </div>
+    </div>
+  </div>
+);
 const item2 = (<p>A Detail List Item to prove float concept when screen readjusts smaller.</p>);
 const accessoryElement = (<p> A sample accessory container </p>);
 
@@ -18,7 +28,7 @@ const DetailViewDividedSmallerTitles = () => (
       <div className={cx('graph-content')}>
         This is where any visualizations would go
       </div>
-)}
+    )}
     accessory={accessoryElement}
     details={[
       (