+++ packages/terra-clinical-result/tests/jest/common/other/KnownNoData.test.jsx
@@ -6,6 +6,7 @@ import KnownNoData from '../../../../src/common/other/_KnownNoData';
 describe('KnownNoData', () => {
   it('should render a default KnownNoData', () => {
     const result = shallowWithIntl(<KnownNoData />).dive();
+    expect(result.find('VisuallyHiddenText').at(0).prop('text')).toEqual('Terra.clinicalResult.resultNoData');
     expect(result).toMatchSnapshot();
   });
 });